+++ packages/spacecat-shared-data-access/src/models/audit/audit.model.js
@@ -41,6 +41,8 @@ class Audit extends BaseModel {
     REDIRECT_CHAINS: 'redirect-chains',
     BROKEN_BACKLINKS: 'broken-backlinks',
     BROKEN_INTERNAL_LINKS: 'broken-internal-links',
+    CONTENT_FRAGMENT_UNUSED: 'content-fragment-unused',
+    CONTENT_FRAGMENT_UNUSED_AUTO_FIX: 'content-fragment-unused-auto-fix',
     EXPERIMENTATION: 'experimentation',
     CONVERSION: 'conversion',
     ORGANIC_KEYWORDS: 'organic-keywords',
@@ -79,6 +81,7 @@ class Audit extends BaseModel {
     LLMO_REFERRAL_TRAFFIC: 'llmo-referral-traffic',
     PAGE_INTENT: 'page-intent',
     NO_CTA_ABOVE_THE_FOLD: 'no-cta-above-the-fold',
+    TOC: 'toc',
   };
 
   static AUDIT_TYPE_PROPERTIES = {